requirements.txt
umap-learn
numpy
scikit-learn
gensim
